projects/sandbox/train/pyproject.toml
[tool.poetry]
name = "train"
version = "0.1.0"
description = "Train aframe"
authors = ["ethan.marx <ethan.marx@ligo.org>"]
license = "MIT"
#readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
torchmetrics = "^0.11"
ml4gw = {path = "../../../ml4gw", develop = true}
"aframe.logging" = {path = "../../../libs/logging", develop = true}
"aframe.trainer" = {path = "../../../libs/trainer", develop = true}

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu116"
secondary = true
default = false

[tool.poetry.group.dev.dependencies]
pytest = "^7.3"
jupyter = "^1.0.0"

[tool.poetry.scripts]
train = "train.train:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
